Simply civicrm handler naming
